Wrap overflow view in modal
