+++ components/newListing/Address/AutoComplete/index.js
@@ -1,84 +1,81 @@
-import _ from 'lodash'
 import React, {PureComponent} from 'react'
-import {View} from 'react-native'
-import {GooglePlacesAutocomplete} from 'react-native-google-places-autocomplete'
+import {View, TouchableWithoutFeedback, Modal} from 'react-native'
 
 import {required} from '@/lib/validations'
-import {GOOGLE_PLACES_API_KEY} from '@/lib/config'
 import {field} from '@/components/shared/Form/Field'
+import Text from '@/components/shared/Text'
+import AutoCompleteModal from './Modal'
 import styles from './styles'
 
 @field({validations: [required('O endereço é obrigatório')]})
-export default class AddressAutoComplete extends PureComponent {
-  static defaultProps = {
-    value: {}
-  }
-
+export default class AutoComplete extends PureComponent {
   state = {
-    selection: {start: 0, end: 0}
+    active: false,
+    value: null,
+    layout: null
   }
 
+  inputButton = React.createRef()
   autoComplete = React.createRef()
 
-  componentDidUpdate(prev) {
-    if (!_.isEqual(prev.value, this.props.value) && this.props.value) {
-      const {street, streetNumber, secondaryAddress} = this.props.value
-      let text = street
-      if (streetNumber) text += ', ' + streetNumber
-      if (secondaryAddress) text += ' - ' + secondaryAddress
-      this.autoComplete.current.setAddressText(text)
-    }
+  onShowModal = () =>
+    this.inputButton.current.measure((_, __, width, height, pageX, pageY) =>
+      this.setState(
+        {
+          active: true,
+          layout: {
+            width,
+            height,
+            pageX,
+            pageY
+          }
+        },
+        () => {
+          requestAnimationFrame(() => this.autoComplete.current.input.focus())
+        }
+      )
+    )
+
+  onHideModal = () => this.setState({active: false})
+
+  onChangeText = (value) => {
+    this.setState({value})
   }
 
-  onSelectionChange = ({nativeEvent: {selection}}) => this.setState({selection})
-
-  onChange = (place) => {
-    const {onChange} = this.props
-    const streetAddress = place.structured_formatting.main_text.split(',')
-    const secondaryAddress = place.structured_formatting.secondary_text
-    if (!streetAddress[1]) {
-      const start = streetAddress[0].length + 2
-      this.setState({
-        selection: {
-          start,
-          end: start + 6
-        }
-      })
-      requestAnimationFrame(() =>
-        this.autoComplete.current.refs.textInput.focus()
-      )
-      streetAddress[1] = 'número'
+  get modalLayout() {
+    const {layout} = this.state
+    if (!layout) return undefined
+    const {width, pageX, pageY} = layout
+    return {
+      width,
+      left: pageX,
+      top: pageY,
+      flexGrow: 1
     }
-    onChange({
-      street: streetAddress[0],
-      streetNumber: streetAddress[1],
-      secondaryAddress
-    })
   }
 
   render() {
+    const {active, value} = this.state
+
     return (
-      <GooglePlacesAutocomplete
-        fetchDefails
-        ref={this.autoComplete}
-        autoFocus={false}
-        horizontal={false}
-        enablePoweredByContainer={false}
-        styles={styles}
-        minLength={3}
-        getDefaultValue={() => ''}
-        onPress={this.onChange}
-        filterReverseGeocodingByTypes={['geocode', 'street_address']}
-        query={{
-          key: GOOGLE_PLACES_API_KEY,
-          language: 'pt-BR',
-          types: 'address'
-        }}
-        textInputProps={{
-          onSelectionChange: this.onSelectionChange,
-          selection: this.state.selection
-        }}
-      />
+      <View>
+        <TouchableWithoutFeedback onPress={this.onShowModal}>
+          <View ref={this.inputButton} style={styles.textInputContainer}>
+            <Text style={styles.textInput}>{value}</Text>
+          </View>
+        </TouchableWithoutFeedback>
+        <Modal transparent visible={active} onRequestClose={this.onHideModal}>
+          <TouchableWithoutFeedback onPress={this.onHideModal}>
+            <View style={this.modalLayout}>
+              <AutoCompleteModal
+                ref={this.autoComplete}
+                onChangeText={this.onChangeText}
+                {...this.props}
+              />
+            </View>
+          </TouchableWithoutFeedback>
+        </Modal>
+      </View>
     )
   }
 }
